requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.12.1
distro==1.9.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.7
imbalanced-learn==0.12.3
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==2.0.0
openai==1.42.0
opencv-python==4.10.0.84
oscrypto==1.3.0
packaging==24.1
pandas==2.2.2
pillow==10.0.0
protobuf==5.27.2
pyarrow==17.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyHanko==0.24.0
pyhanko-certvalidator==0.26.3
pyparsing==3.1.2
pypdf==4.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.35.1
regex==2024.7.24
reportlab==4.0.9
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
setuptools==74.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.36.0
svglib==1.5.1
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.4.0
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
xhtml2pdf==0.2.15
streamlit
transformers
